requirements.txt
amqp==5.3.1
asgiref==3.7.2
attrs==25.3.0
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreschema==0.0.4
cssselect2==0.8.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
fonttools==4.60.1
gunicorn==23.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Markdown==3.5.1
MarkupSafe==2.1.3
mysqlclient==2.2.5
openapi-codec==1.3.2
packaging==24.1
pillow==11.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.8.0
pyphen==0.17.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
setuptools==80.9.0
simplejson==3.19.2
six==1.17.0
sqlparse==0.4.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
reportlab
